Correction chap 1

+++ Algorithmic_Polytech/1-Complexite/code/puissance.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
+#include <stdlib.h>
 
-double puissance(int a, int n) { 
+double puissance(double a, int n) {
     int i;
     double r = 1;
     for(i = 0; i < n; i++) {
@@ -10,10 +11,10 @@ double puissance(int a, int n) {
 }
 
 int main(int argc, char *argv[]) {
-    int a = 1;
+    double a = 1;
     int n = 1;
 
-    if(argc > 1) a = atoi(argv[1]);
+    if(argc > 1) a = atof(argv[1]);
     if(argc > 2) n = atoi(argv[2]);
 
     printf("%f\n",puissance(a,n));
